agents/requirements.txt
# AutoGen Design Debate System Dependencies
autogen>=0.4.0
google-generativeai>=0.8.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0
pydantic>=2.5.0
httpx>=0.26.0
groq>=0.11.0
Pillow>=10.0.0
aiofiles>=23.2.0
